+++ backend/src/models/onboarding.py
@@ -1,8 +1,9 @@
 from __future__ import annotations
 
-from typing import List, Optional
+from datetime import date
+from typing import List, Literal, Optional
 
-from pydantic import BaseModel, Field
+from pydantic import BaseModel, EmailStr, Field
 
 
 class OnboardingAssignmentDto(BaseModel):
@@ -79,4 +80,183 @@ class OnboardingMetricValuesResponseDto(BaseModel):
     savedItemCount: int = 0
     message: str = "OK"
     implementationStatus: str = "READY"
+
+
+class OnboardingAssignmentBulkAssignRequestDto(BaseModel):
+    companyId: int
+    reportingYear: int
+    cycleType: str = "PRE_DMA_G0"
+    metricIds: List[str] = Field(..., min_length=1)
+    assigneeEmail: EmailStr
+    dueDate: Optional[date] = None
+    sendInviteYn: bool = True
+
+
+class OnboardingAssignmentBulkUnassignRequestDto(BaseModel):
+    companyId: int
+    reportingYear: int
+    cycleType: str = "PRE_DMA_G0"
+    metricIds: List[str] = Field(..., min_length=1)
+
+
+class OnboardingAssignmentPatchRequestDto(BaseModel):
+    companyId: int
+    reportingYear: int
+    cycleType: str = "PRE_DMA_G0"
+    assigneeEmail: EmailStr
+    dueDate: Optional[date] = None
+    sendInviteYn: bool = True
+
+
+class OnboardingAssignmentBulkAssignResponseDto(BaseModel):
+    success: bool = True
+    companyId: int
+    reportingYear: int
+    cycleId: int
+    metricIds: List[str]
+    assignmentCount: int
+    assignmentStatus: str
+    assigneeResolvedYn: bool
+    inviteCreatedYn: bool = False
+    inviteReusedYn: bool = False
+    inviteId: Optional[int] = None
+    mailQueuedYn: bool = False
+    warning: Optional[str] = None
+
+
+class OnboardingAssignmentBulkUnassignResponseDto(BaseModel):
+    success: bool = True
+    companyId: int
+    reportingYear: int
+    cycleId: int
+    metricIds: List[str]
+    unassignedCount: int
+    revokedInviteIds: List[int] = Field(default_factory=list)
+    warning: Optional[str] = None
+
+
+class OnboardingAssignmentItemDto(BaseModel):
+    metricId: str
+    metricName: Optional[str] = None
+    assignmentStatus: str
+    assigneeUserId: Optional[int] = None
+    assigneeEmailMasked: Optional[str] = None
+    dueDate: Optional[str] = None
+    inviteId: Optional[int] = None
+    inviteStatus: Optional[str] = None
+
+
+class OnboardingAssignmentListResponseDto(BaseModel):
+    success: bool = True
+    companyId: int
+    reportingYear: int
+    cycleId: Optional[int] = None
+    cycleType: str
+    items: List[OnboardingAssignmentItemDto] = Field(default_factory=list)
+
+
+class OnboardingAssignmentDetailResponseDto(BaseModel):
+    success: bool = True
+    companyId: int
+    reportingYear: int
+    cycleId: Optional[int] = None
+    cycleType: str
+    item: OnboardingAssignmentItemDto
+
+
+ApprovalActionStatus = Literal[
+    "NOT_STARTED",
+    "DRAFT",
+    "SUBMITTED",
+    "REVIEWED",
+    "APPROVED",
+    "REJECTED",
+]
+
+
+class OnboardingApprovalRequestDto(BaseModel):
+    companyId: int
+    reportingYear: int
+    metricId: str = "G0-02"
+
+
+class OnboardingApprovalDecisionRequestDto(OnboardingApprovalRequestDto):
+    commentText: Optional[str] = None
+
+
+class OnboardingApprovalItemDto(BaseModel):
+    companyId: int
+    reportingYear: int
+    metricId: str
+    metricName: Optional[str] = None
+    approvalStatus: ApprovalActionStatus
+    inputUserId: Optional[int] = None
+    assigneeUserId: Optional[int] = None
+    cycleId: Optional[int] = None
+    assignmentId: Optional[int] = None
+    requiredAtomicCount: int = 0
+    completedAtomicCount: int = 0
+    submittedAtomicCount: int = 0
+    approvedAtomicCount: int = 0
+    missingAtomicMetricIds: List[str] = Field(default_factory=list)
+    submittedAt: Optional[str] = None
+    approvedAt: Optional[str] = None
+    commentText: Optional[str] = None
+    selfSubmittedYn: bool = False
+
+
+class OnboardingApprovalListDataDto(BaseModel):
+    items: List[OnboardingApprovalItemDto] = Field(default_factory=list)
+
+
+class OnboardingApprovalListResponseDto(BaseModel):
+    success: bool = True
+    data: OnboardingApprovalListDataDto
+
+
+class OnboardingApprovalStatusDataDto(OnboardingApprovalItemDto):
+    rollupReadyYn: bool = False
+
+
+class OnboardingApprovalStatusResponseDto(BaseModel):
+    success: bool = True
+    data: OnboardingApprovalStatusDataDto
+
+
+class OnboardingApprovalActionResponseDto(BaseModel):
+    success: bool = True
+    data: OnboardingApprovalStatusDataDto
+    message: str = "OK"
+
+
+class OnboardingInviteListItemDto(BaseModel):
+    inviteId: int
+    invitePublicId: str
+    inviteEmailMasked: Optional[str] = None
+    targetRoleCode: str
+    inviteStatus: str
+    expiresAt: Optional[str] = None
+    lastSentAt: Optional[str] = None
+    resendCount: int = 0
+    assignedMetricIds: List[str] = Field(default_factory=list)
+
+
+class OnboardingInviteListResponseDto(BaseModel):
+    success: bool = True
+    companyId: int
+    cycleId: Optional[int] = None
+    items: List[OnboardingInviteListItemDto] = Field(default_factory=list)
+
+
+class OnboardingInviteCompanyRequestDto(BaseModel):
+    companyId: int
+
+
+class OnboardingInviteActionResponseDto(BaseModel):
+    success: bool = True
+    companyId: int
+    inviteId: int
+    inviteStatus: str
+    mailQueuedYn: bool = False
+    warning: Optional[str] = None
