--- backend/src/services/onboarding_approvals/__init__.py
@@ -1,2 +1,0 @@
-"""Onboarding approval service package."""
-
