--- backend/src/services/onboarding_assignments/__init__.py
@@ -1,2 +1,0 @@
-"""Onboarding assignment service package."""
-
