--- backend/src/services/onboarding_invites/__init__.py
@@ -1,1 +1,0 @@
-"""Onboarding invite service package."""
