+++ backend/src/services/onboardings/service.py
@@ -1,23 +1,54 @@
-from __future__ import annotations
+﻿from __future__ import annotations
 
 from typing import Optional
 
 from src.models.onboarding import (
+    OnboardingApprovalActionResponseDto,
+    OnboardingApprovalItemDto,
+    OnboardingApprovalListDataDto,
+    OnboardingApprovalListResponseDto,
+    OnboardingApprovalRequestDto,
+    OnboardingApprovalStatusDataDto,
+    OnboardingApprovalStatusResponseDto,
     OnboardingAssignmentDto,
+    OnboardingAssignmentBulkAssignRequestDto,
+    OnboardingAssignmentBulkAssignResponseDto,
+    OnboardingAssignmentBulkUnassignRequestDto,
+    OnboardingAssignmentBulkUnassignResponseDto,
+    OnboardingAssignmentDetailResponseDto,
+    OnboardingAssignmentItemDto,
+    OnboardingAssignmentListResponseDto,
+    OnboardingAssignmentPatchRequestDto,
     OnboardingAtomicItemDto,
+    OnboardingInviteActionResponseDto,
+    OnboardingInviteListItemDto,
+    OnboardingInviteListResponseDto,
     OnboardingMetricItemDto,
     OnboardingMetricsResponseDto,
     OnboardingMetricValuesRequestDto,
     OnboardingMetricValuesResponseDto,
 )
 from src.utils import onboardingrepository as repo
+from src.utils.companyscope import checkScope
 
 
 CYCLE_TYPE_PRE_DMA_G0 = "PRE_DMA_G0"
-PRE_DMA_G0_CYCLE_NOT_READY = "PRE_DMA_G0_CYCLE_NOT_READY: 보고서 발행 기준을 먼저 선택해 주세요."
+PRE_DMA_G0_CYCLE_NOT_READY = "PRE_DMA_G0_CYCLE_NOT_READY: 보고연도 프로젝트를 먼저 시작해 주세요."
 PRE_DMA_G0_SCOPE_NOT_READY = "PRE_DMA_G0_SCOPE_NOT_READY: 온보딩 지표 범위가 초기화되지 않았습니다. 기존 프로젝트를 재개해 주세요."
 STRUCTURED_LOOKUP_IDS = {"G0-05__QL0002", "G0-06__QL0001"}
 EDITABLE_INPUT_MODES = {"MANUAL_NUMBER", "MANUAL_TEXTAREA", "YEAR_RANGE"}
+SUPPORTED_CYCLE_TYPE = "PRE_DMA_G0"
+SUPPORTED_TARGET_ROLE = "EMPLOYEE"
+INVITE_EXPIRE_DAYS = 7
+ASSIGNMENT_MANAGER_ROLES = {"ADMIN", "ESG"}
+ASSIGNMENT_MANAGER_ROLE_NAMES = {"관리자", "ESG담당자"}
+PRE_DMA_G0_CYCLE_NOT_READY_MESSAGE = "PRE_DMA_G0_CYCLE_NOT_READY: 기존 PRE_DMA_G0 workflow를 먼저 시작해 주세요."
+APPROVER_ROLES = {"ADMIN", "ESG"}
+APPROVER_ROLE_NAMES = {"관리자", "ESG담당자"}
+
+
+class PreDmaG0CycleNotReadyError(ValueError):
+    pass
 
 
 def listMetrics(
@@ -280,3 +311,328 @@ def floatOrNone(value):
         return float(value)
     except (TypeError, ValueError):
         return None
+
+
+def bulkAssign(request: OnboardingAssignmentBulkAssignRequestDto, userModel) -> OnboardingAssignmentBulkAssignResponseDto:
+    checkScope(request.companyId, userModel)
+    checkManager(userModel)
+    checkCycleType(request.cycleType)
+    cycle = requirePreDmaG0Cycle(request.companyId, request.reportingYear)
+    metricIds = repo.validateCycleMetricIds(int(cycle["id"]), request.companyId, request.metricIds)
+    result = repo.bulkAssignMetrics(
+        companyId=request.companyId,
+        reportingYear=request.reportingYear,
+        cycle=cycle,
+        metricIds=metricIds,
+        assigneeEmail=request.assigneeEmail,
+        dueDate=request.dueDate,
+        sendInviteYn=request.sendInviteYn,
+        actorUserId=getActorUserId(userModel),
+    )
+    mailQueuedYn, warning = publishMailEvent(result.get("mailEvent"))
+    return OnboardingAssignmentBulkAssignResponseDto(
+        companyId=result["companyId"],
+        reportingYear=result["reportingYear"],
+        cycleId=result["cycleId"],
+        metricIds=result["metricIds"],
+        assignmentCount=result["assignmentCount"],
+        assignmentStatus=result["assignmentStatus"],
+        assigneeResolvedYn=result["assigneeResolvedYn"],
+        inviteCreatedYn=result["inviteCreatedYn"],
+        inviteReusedYn=result["inviteReusedYn"],
+        inviteId=result.get("inviteId"),
+        mailQueuedYn=mailQueuedYn,
+        warning=warning,
+    )
+
+
+def listAssignmentItems(
+    companyId: int,
+    reportingYear: int,
+    cycleType: str,
+    userModel,
+) -> OnboardingAssignmentListResponseDto:
+    checkScope(companyId, userModel)
+    checkManager(userModel)
+    checkCycleType(cycleType)
+    cycle = repo.resolvePreDmaG0Cycle(companyId=companyId, reportingYear=reportingYear)
+    items = [OnboardingAssignmentItemDto(**item) for item in repo.listAssignments(companyId, reportingYear, cycle)]
+    return OnboardingAssignmentListResponseDto(
+        companyId=companyId,
+        reportingYear=reportingYear,
+        cycleId=int(cycle["id"]) if cycle else None,
+        cycleType=SUPPORTED_CYCLE_TYPE,
+        items=items,
+    )
+
+
+def getAssignmentItem(
+    metricId: str,
+    companyId: int,
+    reportingYear: int,
+    cycleType: str,
+    userModel,
+) -> OnboardingAssignmentDetailResponseDto:
+    checkScope(companyId, userModel)
+    checkManager(userModel)
+    checkCycleType(cycleType)
+    cycle = requirePreDmaG0Cycle(companyId, reportingYear)
+    metricIds = repo.validateCycleMetricIds(int(cycle["id"]), companyId, [metricId])
+    response = OnboardingAssignmentListResponseDto(
+        companyId=companyId,
+        reportingYear=reportingYear,
+        cycleId=int(cycle["id"]),
+        cycleType=SUPPORTED_CYCLE_TYPE,
+        items=[OnboardingAssignmentItemDto(**item) for item in repo.listAssignments(companyId, reportingYear, cycle)],
+    )
+    for item in response.items:
+        if item.metricId == metricIds[0]:
+            return OnboardingAssignmentDetailResponseDto(
+                companyId=companyId,
+                reportingYear=reportingYear,
+                cycleId=response.cycleId,
+                cycleType=response.cycleType,
+                item=item,
+            )
+    raise ValueError(f"metricId was not found: {metricId}")
+
+
+def patchAssignment(metricId: str, request: OnboardingAssignmentPatchRequestDto, userModel) -> OnboardingAssignmentBulkAssignResponseDto:
+    bulkRequest = OnboardingAssignmentBulkAssignRequestDto(
+        companyId=request.companyId,
+        reportingYear=request.reportingYear,
+        cycleType=request.cycleType,
+        metricIds=[metricId],
+        assigneeEmail=request.assigneeEmail,
+        dueDate=request.dueDate,
+        sendInviteYn=request.sendInviteYn,
+    )
+    return bulkAssign(bulkRequest, userModel)
+
+
+def bulkUnassign(request: OnboardingAssignmentBulkUnassignRequestDto, userModel) -> OnboardingAssignmentBulkUnassignResponseDto:
+    checkScope(request.companyId, userModel)
+    checkManager(userModel)
+    checkCycleType(request.cycleType)
+    cycle = requirePreDmaG0Cycle(request.companyId, request.reportingYear)
+    metricIds = repo.validateCycleMetricIds(int(cycle["id"]), request.companyId, request.metricIds)
+    result = repo.bulkUnassignMetrics(
+        companyId=request.companyId,
+        reportingYear=request.reportingYear,
+        cycle=cycle,
+        metricIds=metricIds,
+    )
+    return OnboardingAssignmentBulkUnassignResponseDto(**result)
+
+
+def publishMailEvent(mailEvent: Optional[dict]) -> tuple[bool, Optional[str]]:
+    if not mailEvent:
+        return False, None
+    try:
+        from src.utils.kafkasv import sendToKafka
+
+        sendToKafka(mailEvent)
+        return True, None
+    except Exception as e:
+        return False, f"Mail queue failed: {type(e).__name__}"
+
+
+def requirePreDmaG0Cycle(companyId: int, reportingYear: int) -> dict:
+    cycle = repo.resolvePreDmaG0Cycle(companyId=companyId, reportingYear=reportingYear)
+    if not cycle:
+        raise PreDmaG0CycleNotReadyError(PRE_DMA_G0_CYCLE_NOT_READY_MESSAGE)
+    return cycle
+
+
+def checkCycleType(cycleType: str) -> None:
+    if str(cycleType or "").upper() != repo.CYCLE_TYPE_PRE_DMA_G0:
+        raise ValueError("Only PRE_DMA_G0 cycleType is supported")
+
+
+def checkManager(userModel) -> None:
+    role = str(readUserField(userModel, "role") or "").strip().upper()
+    roleName = str(readUserField(userModel, "role_name") or "").strip()
+    if role in ASSIGNMENT_MANAGER_ROLES or roleName in ASSIGNMENT_MANAGER_ROLE_NAMES:
+        return
+    raise PermissionError("Only ESG담당자 or 관리자 can manage onboarding assignments")
+
+
+def submitApproval(request: OnboardingApprovalRequestDto, userModel) -> OnboardingApprovalActionResponseDto:
+    checkScope(request.companyId, userModel)
+    checkSupportedMetric(request.metricId)
+    summary = repo.submitG002Approval(
+        companyId=request.companyId,
+        reportingYear=request.reportingYear,
+        reportBasisType=None,
+        sourceMaterialityRunId=None,
+        actorUserId=getActorUserId(userModel),
+    )
+    return actionResponse(summary, "Submitted")
+
+
+def approveApproval(request, userModel) -> OnboardingApprovalActionResponseDto:
+    checkScope(request.companyId, userModel)
+    checkSupportedMetric(request.metricId)
+    checkApprover(userModel)
+    summary = repo.approveG002Approval(
+        companyId=request.companyId,
+        reportingYear=request.reportingYear,
+        actorUserId=getActorUserId(userModel),
+        commentText=getattr(request, "commentText", None),
+    )
+    return actionResponse(summary, "Approved")
+
+
+def rejectApproval(request, userModel) -> OnboardingApprovalActionResponseDto:
+    checkScope(request.companyId, userModel)
+    checkSupportedMetric(request.metricId)
+    checkApprover(userModel)
+    commentText = (getattr(request, "commentText", None) or "").strip()
+    if not commentText:
+        raise ValueError("commentText is required for rejection")
+    summary = repo.rejectG002Approval(
+        companyId=request.companyId,
+        reportingYear=request.reportingYear,
+        actorUserId=getActorUserId(userModel),
+        commentText=commentText,
+    )
+    return actionResponse(summary, "Rejected")
+
+
+def listApprovals(
+    companyId: int,
+    reportingYear: Optional[int],
+    status: Optional[str],
+    cycleType: Optional[str],
+    userModel,
+) -> OnboardingApprovalListResponseDto:
+    checkScope(companyId, userModel)
+    if cycleType and str(cycleType).upper() != repo.CYCLE_TYPE_PRE_DMA_G0:
+        return OnboardingApprovalListResponseDto(data=OnboardingApprovalListDataDto(items=[]))
+    items = [
+        itemDto(summary, userModel)
+        for summary in repo.listApprovalSummaries(companyId, reportingYear, status, cycleType)
+    ]
+    return OnboardingApprovalListResponseDto(data=OnboardingApprovalListDataDto(items=items))
+
+
+def getApprovalStatus(
+    companyId: int,
+    reportingYear: int,
+    metricId: str,
+    userModel,
+) -> OnboardingApprovalStatusResponseDto:
+    checkScope(companyId, userModel)
+    checkSupportedMetric(metricId)
+    summary = repo.buildApprovalSummary(companyId, reportingYear, metricId)
+    return OnboardingApprovalStatusResponseDto(data=statusDto(summary, userModel))
+
+
+def ensureWorkflowPreDmaG0Cycle(run: dict, actorUserId: Optional[int] = None) -> dict:
+    if not run:
+        return {}
+    return repo.ensurePreDmaG0Cycle(
+        companyId=int(run["company_id"]),
+        reportingYear=int(run["reporting_year"]),
+        reportBasisType=run.get("report_basis_type"),
+        sourceMaterialityRunId=int(run["id"]) if run.get("id") is not None else None,
+        actorUserId=actorUserId,
+    )
+
+
+def actionResponse(summary: dict, message: str) -> OnboardingApprovalActionResponseDto:
+    return OnboardingApprovalActionResponseDto(
+        data=statusDto(summary, None),
+        message=message,
+    )
+
+
+def itemDto(summary: dict, userModel) -> OnboardingApprovalItemDto:
+    payload = dict(summary)
+    payload.pop("selfSubmittedYn", None)
+    return OnboardingApprovalItemDto(
+        **payload,
+        selfSubmittedYn=checkSelfSubmitted(summary, userModel),
+    )
+
+
+def statusDto(summary: dict, userModel) -> OnboardingApprovalStatusDataDto:
+    payload = dict(summary)
+    payload.pop("selfSubmittedYn", None)
+    return OnboardingApprovalStatusDataDto(
+        **payload,
+        selfSubmittedYn=checkSelfSubmitted(summary, userModel),
+        rollupReadyYn=int(summary.get("approvedAtomicCount") or 0) >= len(repo.REQUIRED_ATOMIC_IDS),
+    )
+
+
+def checkSupportedMetric(metricId: str) -> None:
+    if metricId != repo.METRIC_ID_G0_02:
+        raise ValueError("Only G0-02 approval is supported in MVP")
+
+
+def checkApprover(userModel) -> None:
+    role = str(readUserField(userModel, "role") or "").strip().upper()
+    roleName = str(readUserField(userModel, "role_name") or "").strip()
+    if role in APPROVER_ROLES or roleName in APPROVER_ROLE_NAMES:
+        return
+    raise PermissionError("Only ESG담당자 or 관리자 can approve/reject G0-02 inputs")
+
+
+def checkSelfSubmitted(summary: dict, userModel) -> bool:
+    actorUserId = getActorUserId(userModel)
+    inputUserId = summary.get("inputUserId")
+    return actorUserId is not None and inputUserId is not None and int(actorUserId) == int(inputUserId)
+
+
+def listInviteItems(companyId: int, cycleId: Optional[int], status: Optional[str], userModel) -> OnboardingInviteListResponseDto:
+    checkScope(companyId, userModel)
+    checkManager(userModel)
+    return OnboardingInviteListResponseDto(
+        companyId=companyId,
+        cycleId=cycleId,
+        items=[
+            OnboardingInviteListItemDto(**item)
+            for item in repo.listInvites(companyId, cycleId, status)
+        ],
+    )
+
+
+def resendInviteMail(inviteId: int, companyId: int, userModel) -> OnboardingInviteActionResponseDto:
+    checkScope(companyId, userModel)
+    checkManager(userModel)
+    result = repo.resendInvite(inviteId, companyId)
+    mailQueuedYn, warning = publishMailEvent(result.get("mailEvent"))
+    return OnboardingInviteActionResponseDto(
+        companyId=companyId,
+        inviteId=inviteId,
+        inviteStatus=result["inviteStatus"],
+        mailQueuedYn=mailQueuedYn,
+        warning=warning,
+    )
+
+
+def revokeInviteMail(inviteId: int, companyId: int, userModel) -> OnboardingInviteActionResponseDto:
+    checkScope(companyId, userModel)
+    checkManager(userModel)
+    result = repo.revokeInvite(inviteId, companyId)
+    return OnboardingInviteActionResponseDto(
+        companyId=companyId,
+        inviteId=inviteId,
+        inviteStatus=result["inviteStatus"],
+        mailQueuedYn=False,
+        warning=None,
+    )
+
+
+def getActorUserId(userModel) -> Optional[int]:
+    value = readUserField(userModel, "id")
+    return int(value) if value is not None else None
+
+
+def readUserField(userModel, key: str):
+    if userModel is None:
+        return None
+    if isinstance(userModel, dict):
+        return userModel.get(key)
+    return getattr(userModel, key, None)
